fix/analyzeAst
