+++ package.json
@@ -19,7 +19,7 @@
   "dependencies": {
     "@types/typescript": "^2.0.0",
     "acorn": "^8.11.3",
-    "acorn-walk": "^8.3.2",
-    "typescript": "^5.4.5"
+    "acorn-typescript": "^1.4.13",
+    "acorn-walk": "^8.3.2"
   }
 }
